Refactored Python's socket async client to separate CME and CMD clients, with a base client they extend
