+++ python/python/pybushka/__init__.py
@@ -3,8 +3,7 @@ from pybushka.config import AddressInfo, ClientConfiguration
 from pybushka.constants import OK
 from pybushka.Logger import Level as LogLevel
 from pybushka.Logger import Logger, set_logger_config
-from pybushka.redis_client import RedisClient
-from pybushka.redis_cluster_client import RedisClusterClient
+from pybushka.redis_client import RedisClient, RedisClusterClient
 
 __all__ = [
     "AddressInfo",
